requirements.txt
discord.py[voice]>=2.0.0
python-dotenv
aiohttp>=3.8.0